app/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-appconf==1.0.4
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-select2==7.2.4
django-simple-history==2.8.0
django-userforeignkey==0.3.0
djangorestframework==3.11.2
gunicorn==20.1.0
idna==3.2
kiwisolver==1.3.1
Markdown==3.0.1
matplotlib==3.4.3
mpltern==0.3.3
mysqlclient==1.4.6
numpy==1.19.5
oauthlib==3.1.1
pandas==1.1.5
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2018.9
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
